apps/order-processor/requirements.txt
Flask
python-dotenv
requests
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp